+++ src/Udyam/components/HomePage/HomePage.js
@@ -1,204 +1,12 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import './HomePage.css';
-import ThreeGlobe from 'three-globe';
-import { WebGLRenderer, Scene } from 'three';
-import {
-  PerspectiveCamera,
-  AmbientLight,
-  // DirectionalLight,
-  Color,
-  Fog,
-  // AxesHelper,
-  // DirectionalLightHelper,
-  // CameraHelper,
-  PointLight
-  // SphereGeometry
-} from 'three';
-import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls.js';
-// import { createGlowMesh } from 'three-glow-mesh';
-import countries from './files/globe-data-min.json';
-import travelHistory from './files/my-flights.json';
-// import airportHistory from './files/my-airports.json';
-var renderer, camera, scene, controls;
-// let mouseX = 0;
-// let mouseY = 0;
-// let windowHalfX = window.innerWidth / 2;
-// let windowHalfY = window.innerHeight / 2;
-var Globe;
+import Globe from '../globe/Globe';
 
 function HomePage() {
-  useEffect(() => {
-    init();
-    initGlobe();
-    onWindowResize();
-    animate();
-
-    // SECTION Initializing core ThreeJS elements
-    function init() {
-      // Initialize renderer
-      renderer = new WebGLRenderer({ antialias: true, alpha: true, canvas: document.querySelector('canvas') });
-      renderer.setPixelRatio(window.devicePixelRatio);
-      renderer.setSize(window.innerWidth, window.innerHeight);
-      renderer.setClearColor(0x000000, 0);
-      // renderer.outputEncoding = THREE.sRGBEncoding;
-      // document.body.appendChild(renderer.domElement);
-
-      // Initialize scene, light
-      scene = new Scene();
-      scene.add(new AmbientLight(0xbbbbbb, 0.3));
-      // scene.background = new Color(0x040d21);
-
-      // Initialize camera, light
-      camera = new PerspectiveCamera();
-      camera.aspect = window.innerWidth / window.innerHeight;
-      camera.updateProjectionMatrix();
-
-      var dLight = new PointLight(0xffffff, 0.5);
-      dLight.position.set(-500, 500, 800);
-      camera.add(dLight);
-
-      // var dLight1 = new DirectionalLight(0x7982f6, 1);
-      // dLight1.position.set(-200, 500, 200);
-      // camera.add(dLight1);
-
-      var dLight2 = new PointLight(0x16537e, 0.5);
-      dLight2.position.set(-200, 500, 200);
-      camera.add(dLight2);
-
-      camera.position.z = 210;
-      camera.position.x = -100;
-      camera.position.y = 0;
-
-      scene.add(camera);
-
-      // Additional effects
-      scene.fog = new Fog(0x535ef3, 400, 2000);
-
-      // Helpers
-      // const axesHelper = new AxesHelper(800);
-      // scene.add(axesHelper);
-      // var helper = new DirectionalLightHelper(dLight);
-      // scene.add(helper);
-      // var helperCamera = new CameraHelper(dLight.shadow.camera);
-      // scene.add(helperCamera);
-
-      // Initialize controls
-      controls = new OrbitControls(camera, renderer.domElement);
-      controls.enableDamping = false;
-      controls.dynamicDampingFactor = 0.01;
-      controls.enablePan = false;
-      controls.minDistance = 210;
-      controls.maxDistance = 210;
-      controls.rotateSpeed = 0.3;
-      // controls.zoomSpeed = 0;
-      controls.autoRotate = true;
-
-      controls.minPolarAngle = Math.PI / 3.5;
-      controls.maxPolarAngle = Math.PI - Math.PI / 3;
-
-      window.addEventListener('resize', onWindowResize, false);
-      // document.addEventListener("mousemove", onMouseMove);
-    }
-
-    // SECTION Globe
-    function initGlobe() {
-      // Initialize the Globe
-      Globe = new ThreeGlobe({
-        waitForGlobeReady: true,
-        animateIn: true
-      })
-        .hexPolygonsData(countries.features)
-        .hexPolygonResolution(4)
-        .hexPolygonMargin(0.7)
-        .showAtmosphere(true)
-        .atmosphereColor('#16537e')
-        .atmosphereAltitude(0.2)
-        .hexPolygonColor(() => {
-          return 'rgba(255,255,255, 1)';
-        });
-
-      // NOTE Arc animations are followed after the globe enters the scene
-      setTimeout(() => {
-        Globe.arcsData(travelHistory.flights)
-          .arcColor((e) => {
-            return e.status ? '#9cff00' : '#9cff00';
-          })
-          .arcAltitude((e) => {
-            return e.arcAlt;
-          })
-          .arcStroke((e) => {
-            return e.status ? 0.5 : 0.3;
-          })
-          .arcDashLength(0.9)
-          .arcDashGap(4)
-          .arcDashAnimateTime(2000)
-          .arcsTransitionDuration(2000)
-          .arcDashInitialGap((e) => e.order * 1)
-          // .labelsData(airportHistory.airports)
-          // .labelColor(() => "#ffcb21")
-          // .labelDotOrientation((e) => {
-          //   return e.text === "ALA" ? "top" : "right";
-          // })
-          .labelDotRadius(0.3)
-          .labelSize((e) => e.size)
-          .labelText('city')
-          .labelResolution(6)
-          // .labelAltitude(0.01)
-          // .pointsData(airportHistory.airports)
-          .pointColor(() => '#ffffff')
-          .pointsMerge(true)
-          .pointAltitude(0.07)
-          .pointRadius(0.05);
-      }, 1000);
-
-      // Globe.rotateX(100);
-      // Globe.rotateY(-Math.PI * (5 / 9));
-      // Globe.rotateZ(-Math.PI / 6);
-      const globeMaterial = Globe.globeMaterial();
-      // globe colour
-      globeMaterial.color = new Color(0x3a228a);
-      // globe colour
-      globeMaterial.emissive = new Color(0x220038);
-      globeMaterial.emissiveIntensity = 0.1;
-      globeMaterial.shininess = 0.7;
-
-      // NOTE Cool stuff
-      // globeMaterial.wireframe = true;
-
-      scene.add(Globe);
-    }
-
-    // function onMouseMove(event) {
-    //   mouseX = event.clientX - windowHalfX;
-    //   mouseY = event.clientY - windowHalfY;
-    //   // console.log("x: " + mouseX + " y: " + mouseY);
-    // }
-
-    function onWindowResize() {
-      camera.aspect = window.innerWidth / window.innerHeight;
-      camera.updateProjectionMatrix();
-      // windowHalfX = window.innerWidth / 1.5;
-      // windowHalfY = window.innerHeight / 1.5;
-      renderer.setSize(window.innerWidth, window.innerHeight);
-    }
-
-    function animate() {
-      // camera.position.x +=
-      //   Math.abs(mouseX) <= windowHalfX / 2
-      //     ? (mouseX / 2 - camera.position.x) * 0.005
-      //     : 0;
-      // camera.position.y += (-mouseY / 2 - camera.position.y) * 0.005;
-      camera.lookAt(scene.position);
-      controls.update();
-      renderer.render(scene, camera);
-      requestAnimationFrame(animate);
-    }
-  });
-
   return (
     <div className="main-container">
       <h1 className="heading">UDYAM&apos;23</h1>
-      <canvas id="globe"></canvas>
+      <Globe />
     </div>
   );
 }
