--- src/Udyam/components/HomePage/files/my-flights.json
@@ -1,272 +1,0 @@
-{
-    "type": "FlightsCollection",
-    "flights": [
-        {
-            "type": "flight",
-            "order": 1,
-            "from" : "FRU",
-            "to": "TAS",
-            "flightCode": "HY 778",
-            "date": "Feb 14, 2019",
-            "status": true,
-            "startLat": 42.85,
-            "startLng": 74.583333333,
-            "endLat": 41.266666666667,
-            "endLng": 69.366666666667,
-            "arcAlt": 0.05
-        },
-        {
-            "type": "flight",
-            "order": 2,
-            "from" : "TAS",
-            "to": "ICN",
-            "flightCode": "HY 551",
-            "date": "Feb 15, 2019",
-            "status": true,
-            "startLat": 41.266666666667,
-            "startLng": 69.366666666667,
-            "endLat": 37.472222222222,
-            "endLng": 126.60833333333,
-            "arcAlt": 0.2 
-        },
-        {
-            "type": "flight",
-            "order": 3,
-            "from" : "ICN",
-            "to": "KUL",
-            "flightCode": "D7 507",
-            "date": "Aug 30, 2019",
-            "status": true,
-            "startLat": 37.472222222222,
-            "startLng": 126.60833333333,
-            "endLat": 3.116666666666,
-            "endLng": 101.7,
-            "arcAlt": 0.2 
-        },
-        {
-            "type": "flight",
-            "order": 4,
-            "from" : "KUL",
-            "to": "SIN",
-            "flightCode": "TR 453",
-            "date": "Oct 15, 2019",
-            "status": true,
-            "startLat": 3.11666666666,
-            "startLng": 101.7,
-            "endLat": 1.2916666666667,
-            "endLng": 103.85,
-            "arcAlt": 0.04 
-        },
-        {
-            "type": "flight",
-            "order": 4.5,
-            "from" : "KUL",
-            "to": "SIN",
-            "flightCode": "TR 453",
-            "date": "Oct 22, 2019",
-            "status": true,
-            "startLat": 3.11666666666,
-            "startLng": 101.7,
-            "endLat": 1.2916666666667,
-            "endLng": 103.85,
-            "arcAlt": 0.07 
-        },
-        {
-            "type": "flight",
-            "order": 5,
-            "from" : "KUL",
-            "to": "DPS",
-            "flightCode": "OD 177",
-            "date": "Oct 21, 2019",
-            "status": true,
-            "startLat": 3.11666666666,
-            "startLng": 101.7,
-            "endLat": -8.75,
-            "endLng": 115.16666666667,
-            "arcAlt": 0.1 
-        },
-        {
-            "type": "flight",
-            "order": 6,
-            "from" : "DPS",
-            "to": "KUL",
-            "flightCode": "OD 325",
-            "date": "Oct 31, 2019",
-            "status": true,
-            "startLat": -8.75,
-            "startLng": 115.16666666667,
-            "endLat": 3.116666666666,
-            "endLng": 101.7,
-            "arcAlt": 0.15 
-        },
-        {
-            "type": "flight",
-            "order": 7,
-            "from" : "KUL",
-            "to": "DMK",
-            "flightCode": "AK 882",
-            "date": "Nov 25, 2019",
-            "status": true,
-            "startLat": 3.11666666666,
-            "startLng": 101.7,
-            "endLat": 13.9255,
-            "endLng": 100.5923,
-            "arcAlt": 0.1 
-        },
-        {
-            "type": "flight",
-            "order": 8,
-            "from" : "BKK",
-            "to": "KJA",
-            "flightCode": "S7 6406",
-            "date": "Dec 4, 2019",
-            "status": true,
-            "startLat": 13.6900,
-            "startLng": 100.7501,
-            "endLat": 56.05,
-            "endLng": 92.9,
-            "arcAlt": 0.25 
-        },
-        {
-            "type": "flight",
-            "order": 9,
-            "from" : "KJA",
-            "to": "OVB",
-            "flightCode": "S7 5306",
-            "date": "Dec 5, 2019",
-            "status": true,
-            "startLat": 56.05,
-            "startLng": 92.9,
-            "endLat": 55.033333333333,
-            "endLng": 83,
-            "arcAlt": 0.05 
-        },
-        {
-            "type": "flight",
-            "order": 10,
-            "from" : "OVB",
-            "to": "FRU",
-            "flightCode": "S7 5541",
-            "date": "Dec 5, 2019",
-            "status": true,
-            "startLat": 55.033333333333,
-            "startLng": 83,
-            "endLat": 42.85,
-            "endLng": 74.583333333,
-            "arcAlt": 0.1 
-        },
-        {
-            "type": "flight",
-            "order": 10.5,
-            "from" : "FRU",
-            "to": "ALA",
-            "flightCode": "KC 110",
-            "date": "Mar 8, 2020",
-            "status": true,
-            "startLat": 42.85,
-            "startLng": 74.583333333,
-            "endLat": 43.35,
-            "endLng": 77.033333333333,
-            "arcAlt": 0.08 
-        },
-        {
-            "type": "flight",
-            "order": 11,
-            "from" : "ALA",
-            "to": "ICN",
-            "flightCode": "KC 909",
-            "date": "Mar 9, 2020",
-            "status": true,
-            "startLat": 43.35,
-            "startLng": 77.033333333333,
-            "endLat": 37.472222222222,
-            "endLng": 126.60833333333,
-            "arcAlt": 0.2 
-        },
-        {
-            "type": "flight",
-            "order": 12,
-            "from" : "FRU",
-            "to": "ALA",
-            "flightCode": "KC 110",
-            "date": "Mar 12, 2020",
-            "status": true,
-            "startLat": 42.85,
-            "startLng": 74.583333333,
-            "endLat": 43.35,
-            "endLng": 77.033333333333,
-            "arcAlt": 0.1 
-        },
-        {
-            "type": "flight",
-            "order": 13,
-            "from" : "ALA",
-            "to": "BKK",
-            "flightCode": "KC 931",
-            "date": "Mar 13, 2020",
-            "status": true,
-            "startLat": 43.35,
-            "startLng": 77.033333333333,
-            "endLat": 13.6900,
-            "endLng": 100.7501,
-            "arcAlt": 0.2
-        },
-        {
-            "type": "flight",
-            "order": 13.5,
-            "from" : "ICN",
-            "to": "DMK",
-            "flightCode": "XJ 703",
-            "date": "Mar 19, 2020",
-            "status": true,
-            "startLat": 37.472222222222,
-            "startLng": 126.60833333333,
-            "endLat": 13.9255,
-            "endLng": 100.5923,
-            "arcAlt": 0.1 
-        },
-        {
-            "type": "flight",
-            "order": 14,
-            "from" : "DMK",
-            "to": "ICN",
-            "flightCode": "XJ 700",
-            "date": "Mar 20, 2020",
-            "status": true,
-            "startLat": 13.9255,
-            "startLng": 100.5923,
-            "endLat": 37.472222222222,
-            "endLng": 126.60833333333,
-            "arcAlt": 0.15 
-        },
-        {
-            "type": "flight",
-            "order": 14.5,
-            "from" : "DMK",
-            "to": "ICN",
-            "flightCode": "XJ 702",
-            "date": "Mar 23, 2020",
-            "status": true,
-            "startLat": 13.9255,
-            "startLng": 100.5923,
-            "endLat": 37.472222222222,
-            "endLng": 126.60833333333,
-            "arcAlt": 0.2 
-        },
-        {
-            "type": "flight",
-            "order": 15,
-            "from" : "BKK",
-            "to": "ICN",
-            "flightCode": "TG 656",
-            "date": "Mar 23, 2020",
-            "status": true,
-            "startLat": 13.6900,
-            "startLng": 100.7501,
-            "endLat": 37.472222222222,
-            "endLng": 126.60833333333,
-            "arcAlt": 0.25 
-        }
-        
-    ]
-}
