+++ src/Udyam/components/globe/Globe.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const eGlobe = () => {
+  return <div></div>;
+};
+
+export default eGlobe;
