+++ src/files/my-airports.json
@@ -1,0 +1,93 @@
+{
+  "type": "AirportsCollection",
+  "airports": [
+    {
+      "text": "FRU",
+      "size": 1.0,
+      "country": "Kyrgyzstan",
+      "city": "Bishkek",
+      "lat": "42.85",
+      "lng": "74.583333333"
+    },
+    {
+      "text": "TAS",
+      "size": 1.0,
+      "country": "Uzbekistan",
+      "city": "Tashkent",
+      "lat": "41.266666666667",
+      "lng": "69.366666666667"
+    },
+    {
+      "text": "ALA",
+      "size": 1.0,
+      "country": "Kazakhstan",
+      "city": "Almaty",
+      "lat": "43.35",
+      "lng": "75.033333333333"
+    },
+    {
+      "text": "ICN",
+      "size": 1.0,
+      "country": "South Korea",
+      "city": "Incheon",
+      "lat": "37.472222222222",
+      "lng": "126.60833333333"
+    },
+    {
+      "text": "KJA",
+      "size": 1.0,
+      "country": "Russia",
+      "city": "Krasnoyarsk",
+      "lat": "56.05",
+      "lng": "92.9"
+    },
+    {
+      "text": "OVB",
+      "size": 1.0,
+      "country": "Russia",
+      "city": "Novosibirsk",
+      "lat": "55.033333333333",
+      "lng": "83"
+    },
+    {
+      "text": "KUL",
+      "size": 1.0,
+      "country": "Malaysia",
+      "city": "Kuala-Lumpur",
+      "lat": "3.11666666666",
+      "lng": "101.7"
+    },
+    {
+      "text": "SIN",
+      "size": 1.0,
+      "country": "Singapore",
+      "city": "Singapore",
+      "lat": "1.2916666666667",
+      "lng": "103.85"
+    },
+    {
+      "text": "DPS",
+      "size": 1.0,
+      "country": "Indonesia",
+      "city": "Denpasar",
+      "lat": "-8.75",
+      "lng": "115.16666666667"
+    },
+    {
+      "text": "DMK",
+      "size": 1.0,
+      "country": "Thailand",
+      "city": "Bangkok",
+      "lat": "13.9255",
+      "lng": "100.5923"
+    },
+    {
+      "text": "BKK",
+      "size": 1.0,
+      "country": "Thailand",
+      "city": "",
+      "lat": "13.6900",
+      "lng": "100.7501"
+    }
+  ]
+}
